Combine product and login details into product-details.js
