+++ src/js/product-details.js
@@ -128,4 +128,40 @@ const canonicalProducts = {
   ],
 };
 
-export default canonicalProducts;
+const canonicalLogins = [
+  {
+    title: 'Ubuntu Advantage',
+    logoUrl: 'https://assets.ubuntu.com/v1/c5cb0f8e-picto-ubuntu.svg',
+    description: 'Enterprise support service to manage and add value to your Ubuntu deployment',
+    login: 'http://support.canonical.com/',
+    signup: 'https://buy.ubuntu.com/',
+  }, {
+    title: 'Landscape',
+    logoUrl: 'https://assets.ubuntu.com/v1/4964d639-landscape-logo.svg',
+    description: 'Manage, deploy, and monitor your Ubuntu servers',
+    login: 'https://landscape.canonical.com/login/authenticate',
+  }, {
+    title: 'Livepatch',
+    logoUrl: 'https://assets.ubuntu.com/v1/3c1de878-canonical-logo.png',
+    description: 'Apply critical kernel security fixes without rebooting',
+    login: 'https://livepatch.canonical.com/',
+    signup: 'https://auth.livepatch.canonical.com/',
+  }, {
+    title: 'Launchpad.net',
+    logoUrl: 'https://assets.ubuntu.com/v1/6b318257-launchpad-icon.png',
+    description: 'The software collaboration platform behind Ubuntu',
+    login: 'https://launchpad.net/+login',
+  }, {
+    title: 'Snapcraft.io',
+    logoUrl: 'https://assets.ubuntu.com/v1/82dc2022-snapcraft-primary-icon.svg',
+    description: 'Manage your snaps and snap store(s)',
+    login: 'https://snapcraft.io/account',
+  }, {
+    title: 'JAAS',
+    logoUrl: 'https://assets.ubuntu.com/v1/31c507a5-logo-juju-icon.svg',
+    description: 'Deploy, configure, scale and operate your software on public and private clouds',
+    login: 'https://jujucharms.com/login/',
+  },
+];
+
+export { canonicalProducts, canonicalLogins };
